refactor: collapse session barrel into session/index.ts
