+++ packages/opencode/src/session/revert.ts
@@ -5,7 +5,7 @@ import { Snapshot } from "../snapshot"
 import { Storage } from "@/storage"
 import { SyncEvent } from "../sync"
 import { Log } from "../util"
-import * as Session from "./session"
+import * as Session from "."
 import { MessageV2 } from "./message-v2"
 import { SessionID, MessageID, PartID } from "./schema"
 import { SessionRunState } from "./run-state"
